@@ -1,10 +1,10 @@
-module megpoid.dev/go/docker-deploy
+module go.megpoid.dev/docker-deploy
 
-go 1.18
+go 1.24
 
 require (
-	github.com/spf13/pflag v1.0.5
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
+	github.com/spf13/pflag v1.0.6
+	gopkg.in/yaml.v3 v3.0.1
 )
 
 require (
